feat(tools): name the zone-free calendar intents
`dateTime.ts` conflated three different questions in one flat namespace: which
calendar day, what time on the clock, and which absolute instant. A caller
reading `extractDate(x)` could not tell whether a zone was involved, and the
answer differed per helper.

Split the two intents that need no zone into their own modules:

- `plainDate.ts` — a calendar day. No clock, no zone. `2026-09-09` is the same
  day read from Auckland or Los Angeles.
- `plainTime.ts` — a wall clock. No day, no zone. `14:00` is not a moment until
  a day and a zone are supplied.
- `dateTimeInternals.ts` — the `Date` plumbing both need (`isDate`,
  `isDateObject`, `zeroPad`), so neither intent module imports the other.

`dateTime.ts` becomes a legacy adapter that re-exports them. 42 exports before,
42 after: none added, none missing, no signature or behaviour change. All 78
internal `@Tools/dateTime` import sites are untouched, as is every published
name. `tools.plainDate` and `tools.plainTime` are added alongside `tools.dateTime`
so call sites can move one at a time.

Two tests guard what would otherwise rot. `dateTime.extractDate` is asserted
IDENTITY-equal to `plainDate.extractDate` — a second implementation would drift
silently rather than fail. And each bundle is asserted to carry no member of a
foreign intent, so the boundary cannot erode by accident.

The names are Temporal's on purpose. When `Temporal.PlainDate` and
`Temporal.PlainTime` become available on every runtime we operate, these bodies
become one-for-one substitutions instead of a rename.
